requirements.txt
boto3
psycopg2-binary
pandas
XlsxWriter
scikit-learn
jupyter
matplotlib
seaborn
reportlab
openpyxl